% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radbrennie.R
\name{rad_brennie}
\alias{rad_brennie}
\title{Rad Brennie by Stoke and Hype}
\usage{
rad_brennie(stoke, hype, dogs_pet = 5, pizza = 2)
}
\arguments{
\item{dogs_pet}{coefficient is the amount of dogs obviously pet in a day default=5}

\item{pizza}{coefficient is the number of whole pizzas in a day default=2}

\item{Stoke}{is the confidence gained for every completed homework assginment at Bren (confidence/hw turned in)}

\item{Hype}{is excitment gained for number of classes taken that quarter (excitment/class)}
}
\value{
rad (R)
}
\description{
This function determines the amount of Rad a Bren Student has
}
\author{
Sidney
}
