% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rad_df.R
\docType{data}
\name{rad_df}
\alias{rad_df}
\title{Rad Data from Bren}
\format{A data frame with 1000 rows and 2 columns
\describe{
\item{stoke}{amount of stoke (s)}
\item{hype}{amount of hype (h)}
}}
\usage{
rad_df
}
\description{
Data from how much hype and stoke exists at Bren
}
\keyword{datasets}
